src/backend/requirements.txt
fastapi
uvicorn[standard]
ray[default]
sqlalchemy
psycopg2-binary
pymongo
kombu
python-dotenv
pydantic